backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy>=2.0.40
pymysql==1.1.1
alembic==1.13.2
pydantic-settings==2.5.2
typing-extensions>=4.13.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.12
